[OPR-216]: Adds SwiftLint
Includes lots of lint fixes.
