+++ Sources/Core/Shared/RepeatedOperation.swift
@@ -8,116 +8,6 @@
 
 import Foundation
 
-func arc4random<T: IntegerLiteralConvertible>(type: T.Type) -> T {
-    var r: T = 0
-    arc4random_buf(&r, Int(sizeof(T)))
-    return r
-}
-
-/**
- Random Failure Generator
-
- This generator will randomly return nil (instead of the
- composed generator's `next()`) according to the probability
- of failure argument.
-
- For example, to simulate 1% failure rate:
-
- ```swift
- let one = RandomFailGenerator(generator, probability: 0.01)
- ```
-*/
-public struct RandomFailGenerator<G: GeneratorType>: GeneratorType {
-
-    private var generator: G
-    private let shouldNotFail: () -> Bool
-
-    /**
-     Initialize the generator with another generator and expected
-     probabilty of failure.
-
-     - parameter: the generator
-     - parameter probability: the expected probability of failure, defaults to 0.1, or 10%
-    */
-    public init(_ generator: G, probability: Double = 0.1) {
-        self.generator = generator
-        self.shouldNotFail = {
-            let r = (Double(arc4random(UInt64)) / Double(UInt64.max))
-            return r > probability
-        }
-    }
-
-    /// GeneratorType implementation
-    public mutating func next() -> G.Element? {
-        guard shouldNotFail() else {
-            return nil
-        }
-        return generator.next()
-    }
-}
-
-struct FibonacciGenerator: GeneratorType {
-    var currentValue = 0, nextValue = 1
-
-    mutating func next() -> Int? {
-        let result = currentValue
-        currentValue = nextValue
-        nextValue += result
-        return result
-    }
-}
-
-struct FiniteGenerator<G: GeneratorType>: GeneratorType {
-
-    private let limit: Int
-    private var generator: G
-
-    var count: Int = 0
-
-    init(_ generator: G, limit: Int = 10) {
-        self.generator = generator
-        self.limit = limit
-    }
-
-    mutating func next() -> G.Element? {
-        guard count < limit else {
-            return nil
-        }
-        count += 1
-        return generator.next()
-    }
-}
-
-struct MapGenerator<G: GeneratorType, T>: GeneratorType {
-    private let transform: G.Element -> T
-    private var generator: G
-
-
-    init(_ generator: G, transform: G.Element -> T) {
-        self.generator = generator
-        self.transform = transform
-    }
-
-    mutating func next() -> T? {
-        return generator.next().map(transform)
-    }
-}
-
-struct TupleGenerator<Primary: GeneratorType, Secondary: GeneratorType>: GeneratorType {
-
-    private var primary: Primary
-    private var secondary: Secondary
-
-    init(primary: Primary, secondary: Secondary) {
-        self.primary = primary
-        self.secondary = secondary
-    }
-
-    mutating func next() -> (Secondary.Element?, Primary.Element)? {
-        return primary.next().map { ( secondary.next(), $0) }
-    }
-}
-
 /**
  Define a strategy for waiting a given time interval. The strategy
  can then create a NSTimeInterval generator. The strategies are:
@@ -169,46 +59,6 @@ public enum WaitStrategy {
 
     internal func generator() -> IntervalGenerator {
         return IntervalGenerator(self)
-    }
-}
-
-struct IntervalGenerator: GeneratorType {
-
-    let strategy: WaitStrategy
-
-    private var count: Int = 0
-    private lazy var fibonacci = FibonacciGenerator()
-
-    init(_ strategy: WaitStrategy) {
-        self.strategy = strategy
-    }
-
-    mutating func next() -> NSTimeInterval? {
-        switch strategy {
-
-        case .Fixed(let period):
-            return period
-
-        case .Random(let (minimum, maximum)):
-            let r = Double(arc4random(UInt64)) / Double(UInt64.max)
-            return (r * (maximum - minimum)) + minimum
-
-        case .Incrementing(let (initial, increment)):
-            let interval = initial + (NSTimeInterval(count) * increment)
-            count += 1
-            return max(0, interval)
-
-        case .Exponential(let (period, maximum)):
-            let interval = period * pow(2.0, Double(count))
-            count += 1
-            return max(0, min(maximum, interval))
-
-        case .Fibonacci(let (period, maximum)):
-            return fibonacci.next().map { fib in
-                let interval = period * NSTimeInterval(fib)
-                return max(0, min(maximum, interval))
-            }
-        }
     }
 }
 
@@ -460,127 +310,9 @@ public class RepeatedOperation<T where T: NSOperation>: GroupOperation {
     */
     public func addConfigureBlock(block: T -> Void) {
         let config = configure
-        configure = { op in
-            config(op)
-            block(op)
-        }
-    }
-}
-
-/**
-
- ### Repeatable
-
- `Repeatable` is a very simple protocol, which your `NSOperation` subclasses
- can conform to. This allows the previous operation to define whether a new
- one should be executed. For this special case, `RepeatedOperation` can be
- initialized like this:
-
- ```swift
- let operation = RepeatedOperation { MyRepeatableOperation() }
- ```
-
- - see: RepeatedOperation
-*/
-public protocol Repeatable {
-
-    /**
-     Implement this funtion to return true if a new
-     instance should be added to a RepeatedOperation.
-
-     - parameter count: the number of instances already executed within
-     the RepeatedOperation.
-     - returns: a Bool, false will end the RepeatedOperation.
-    */
-    func shouldRepeat(count: Int) -> Bool
-}
-
-public class RepeatableGenerator<G: GeneratorType where G.Element: Repeatable>: GeneratorType {
-
-    private var generator: G
-    private var count: Int = 0
-    private var current: G.Element?
-
-    public init(_ generator: G) {
-        self.generator = generator
-    }
-
-    public func next() -> G.Element? {
-        if let current = current {
-            guard current.shouldRepeat(count) else {
-                return nil
-            }
-        }
-        current = generator.next()
-        count += 1
-        return current
-    }
-}
-
-extension RepeatedOperation where T: Repeatable {
-
-    /**
-     Initialize a RepeatedOperation using a closure with NSOperation subclasses
-     which conform to Repeatable. This is the neatest initializer.
-
-     ```swift
-     let operation = RepeatedOperation { MyRepeatableOperation() }
-     ```
-    */
-    public convenience init(maxCount max: Int? = .None, strategy: WaitStrategy = .Fixed(0.1), body: () -> T?) {
-        self.init(maxCount: max, strategy: strategy, generator: RepeatableGenerator(anyGenerator(body)))
-    }
-}
-
-/**
- RepeatableOperation is an Operation subclass which conforms to Repeatable.
-
- It can be used to make an otherwise non-repeatable Operation repeatable. It
- does this by accepting, in addition to the operation instance, a closure
- shouldRepeat. This closure can be used to capture state (such as errors).
-
- When conforming to Repeatable, the closure is executed, passing in the
- current repeat count.
-*/
-public class RepeatableOperation<T: Operation>: Operation, OperationDidFinishObserver, Repeatable {
-
-    let operation: T
-    let shouldRepeatBlock: Int -> Bool
-
-    /**
-     Initialize the RepeatableOperation with an operation and
-     shouldRepeat closure.
-
-     - parameter [unnamed] operation: the operation instance.
-     - parameter shouldRepeat: a closure of type Int -> Bool
-    */
-    public init(_ operation: T, shouldRepeat: Int -> Bool) {
-        self.operation = operation
-        self.shouldRepeatBlock = shouldRepeat
-        super.init()
-        name = "Repeatable<\(operation.operationName)>"
-        addObserver(CancelledObserver { [weak operation] _ in
-            (operation as? Operation)?.cancel()
-        })
-    }
-
-    /// Override implementation of execute
-    public override func execute() {
-        if !cancelled {
-            operation.addObserver(self)
-            produceOperation(operation)
-        }
-    }
-
-    /// Implementation for Repeatable
-    public func shouldRepeat(count: Int) -> Bool {
-        return shouldRepeatBlock(count)
-    }
-
-    /// Implementation for OperationDidFinishObserver
-    public func didFinishOperation(operation: Operation, errors: [ErrorType]) {
-        if self.operation == operation {
-            finish(errors)
+        configure = { operation in
+            config(operation)
+            block(operation)
         }
     }
 }
