+++ Sources/Extras/Health/iOS/HealthCapability.swift
@@ -43,23 +43,23 @@ public struct HealthRequirement {
 public class HealthCapabilityStatus: NSObject, AuthorizationStatusType {
     typealias DictionaryType = Dictionary<String, HKAuthorizationStatus>
 
-    var _dictionary = DictionaryType()
+    var dictionary = DictionaryType()
 
     /// Access the HKAuthorizationStatus by the share identifier.
     public subscript(key: DictionaryType.Key) -> DictionaryType.Value? {
         get {
-            return _dictionary[key]
+            return dictionary[key]
         }
         set(newStatus) {
-            _dictionary[key] = newStatus
+            dictionary[key] = newStatus
         }
     }
 
 
     /// Determine whether the application permissions have been met.
     public func isRequirementMet(requirement: HealthRequirement) -> Bool {
-        if requirement.shareIdentifiers.isSubsetOf(_dictionary.keys) {
-            let statuses = Set(_dictionary.values)
+        if requirement.shareIdentifiers.isSubsetOf(dictionary.keys) {
+            let statuses = Set(dictionary.values)
             if statuses.count == 1 && statuses.first! == .SharingAuthorized {
                 return true
             }
@@ -240,8 +240,8 @@ extension HealthRequirement: Equatable { }
 /**
  HealthRequirement is Equatable
 */
-public func ==(a: HealthRequirement, b: HealthRequirement) -> Bool {
-    return (a.share == b.share) && (a.read == b.read)
+public func == (lhs: HealthRequirement, rhs: HealthRequirement) -> Bool {
+    return (lhs.share == rhs.share) && (lhs.read == rhs.read)
 }
 
 /**
