+++ Tests/Extras/AddressBook/AddressBookTests.swift
@@ -619,11 +619,3 @@ class AddressBookWritablePropertyTests: XCTestCase {
         XCTAssertNotNil(writable.writer)
     }
 }
-
-
-
-
-
-
-
-
