+++ Tests/Features/LocationCapabilityTests.swift
@@ -203,4 +203,3 @@ class LocationCapabilityTests: OperationTests {
         XCTAssertTrue(didComplete)
     }
 }
-
