+++ Tests/Features/RemoteNotificationConditionTests.swift
@@ -74,4 +74,3 @@ class RemoteNotificationConditionTests: OperationTests {
         }
     }
 }
-
